Add eslint and editorconfig for whitespacing and some unused lines + code split
Changes:
- Add eslint and editorcofig for code styling (whitespacing, unused params, undefined var etc.)
- Split big preload code in separate modules.
