+++ package.json
@@ -16,7 +16,8 @@
     "@electron-forge/maker-rpm": "^6.0.0-beta.63",
     "@electron-forge/maker-squirrel": "^6.0.0-beta.63",
     "@electron-forge/maker-zip": "^6.0.0-beta.63",
-    "electron": "^16.0.7"
+    "electron": "^16.0.7",
+    "eslint": "^8.18.0"
   },
   "dependencies": {
     "clipboard-event": "^1.5.0",
