+++ src/main/webBrowser/webBrowserMain.js
@@ -13,13 +13,13 @@ app.commandLine.appendSwitch('disable-features', 'CrossOriginOpenerPolicy');
 function createWebWindow() {
     parent = main.getMainWindow();
 
-    browser = new BrowserWindow({ 
+    browser = new BrowserWindow({
         title: 'UnCover Browser',
         width: 800,
         height: 500,
         minWidth: 500,
         minHeight: 100,
-        parent: parent, 
+        parent: parent,
         show: false,
         autoHideMenuBar: true,
         webPreferences: {
@@ -29,7 +29,7 @@ function createWebWindow() {
             webviewTag: true,
         }
     })
-    
+
     browser.on('closed', () => {
         browser = null;
     })
@@ -49,7 +49,7 @@ function createWebWindow() {
 }
 
 function createNewWindow(url) {
-    const newWindow = new BrowserWindow({ 
+    const newWindow = new BrowserWindow({
         width: 800,
         height: 500,
         parent: parent,
@@ -71,6 +71,6 @@ function openWebBrowser() {
     browser ? browser.show() : createWebWindow();
 }
 
-ipcMain.handle("open-web-browser", (e, ...args) => {
+ipcMain.handle("open-web-browser", () => {
     openWebBrowser();
 });
