+++ src/preload/modules/clipboard.js
@@ -1,0 +1,26 @@
+const { clipboard } = require('electron');
+const clipboardListener = require('clipboard-event');
+
+clipboardListener.startListening();
+
+function removeImageUrlPrefix(dataUrl) {
+    let data = dataUrl.replace(/^data:image\/\w+;base64,/, "");
+    return Buffer.from(data, 'base64');
+}
+
+const contextBridge = {
+    // Clipboard
+    onClipboardChanged: (event, listener) => clipboardListener.on(event, listener),
+    clearClipboard: () => clipboard.clear(),
+    getClipboardAvailableFormats: () => clipboard.availableFormats(),
+    readClipboardText: () => clipboard.readText(),
+    readClipboardImage: () => clipboard.readImage().toDataURL(),
+    removeImageUrlPrefix: (dataUrl) => removeImageUrlPrefix(dataUrl),
+    imageIsEmpty: () => {
+        return clipboard.readImage().isEmpty()
+    },
+}
+
+module.exports = {
+    clipboard: contextBridge
+};
