+++ src/renderer/js/Helpers/convertPathUrl.js
