+++ src/renderer/js/Helpers/pageHandler.js
