+++ src/renderer/js/helpers/pageHandler.js
