+++ src/renderer/js/templates/webComponentTemplate.js
